Post Processing / Player / GameManager
I added post processing effect (bloom), and I added the player into the game. Working on Pawn and Controller, as well as working on the Game Manager. This will keep track of the player's speed, and will increase the player's speed slowly over time. I've also added particle effects to the player. Aesthetic-wise, this looks like a very good game.