Successfully Optimized Game
Even if it's not a lot, it's definitely a significant change over how it ran before.
Learning that you can combine a Composite Collider 2D with a Tilemap Collider 2D, it reduces the amount of physics based by combining all the colliders as 1 singular piece, which help made the game run a little better.

I still get lag, but sometimes it's hardly noticeable.