Made IRange Interface, and it was a huge Success!!!
Any object can implement IRange and have an OnRangeEnter and OnRangeExit by adding the DistanceCheck component to the object, and adding a Transform to it. This is just made for this game, but the great think about doing it this way is it's not physics based, and since it doesn't really on a fixed time stamp, it's always accurate!