Desposit and Withdraw work as should; Cloud Stats not working
Got the Crystal Value to change upon spending or collecting the crystals. However, I realized that the GetStats code wasn't really working. As I went to check on the debug.log to see what what going on, I was basically right on the money, as I probably coded something in wrong, or formatted it the wrong way, or there's no matching values for it to connect itself to. (Or something like that).